+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 0.1.0
+## [Unreleased]
+
+## [0.1.0] (2019-07-21)
 
 ### Added
-- 
+- Create a new blog with the `blognow [name]` command
+- Build a static site with `blognow`
+- Blog title and tagline
+- Single post page
+- Archive page to view all posts
+
+[Unreleased]: https://github.com/johnjago/blognow/compare/0.1.0...HEAD
+[0.1.0]: https://github.com/johnjago.com/blognow/releases/tag/0.1.0